@@ -1,7 +1,7 @@
 {
   "name": "nek-ui",
   "description": "UI components based on RegularJS",
-  "version": "0.6.1",
+  "version": "0.6.2",
   "main": "./dist/js/nek-ui.js",
   "homepage": "https://nek-ui.kaolafed.com/",
   "license": "MIT",
@@ -17,9 +17,11 @@
   "scripts": {
     "release": "npm run build && ./mkzip.sh && np --yolo",
     "peta": "npm run build && peta",
-    "build": "gulp dist",
-    "dev": "gulp watch --dev",
-    "doc-dev": "gulp watch-doc --dev",
+    "build": "gulp --gulpfile build/gulpfile.js dist",
+    "dev": "gulp --gulpfile build/gulpfile.js watch --dev",
+    "new": "gulp --gulpfile build/gulpfile.js new --dev",
+    "dev1": "webpack-dev-server --open --inline --hot --compress --history-api-fallback --port 8083 --config build/webpack.dev.conf.js",
+    "doc-dev": "gulp --gulpfile build/gulpfile.jswatch-doc --dev",
     "lint": "eslint ./src/js",
     "lint:fix": "eslint --fix ./src/js",
     "precommit": "lint-staged"
@@ -35,11 +37,16 @@
     "bowser": "^1.5.0",
     "clipboard": "^1.6.0",
     "dom-align": "^1.5.2",
-    "moment": "^2.17.1",
+    "js-calendar": "^1.2.3",
+    "moment": "^2.22.2",
+    "popper.js": "^1.14.3",
+    "postcss-salad": "^2.0.1",
+    "regular-state": "^0.6.0-beta.8",
     "regularjs": "0.4.3",
     "reqwest": "^2.0.5",
     "scroll-into-view": "^1.9.1",
-    "validator": "^6.1.0"
+    "validator": "^6.1.0",
+    "resize-observer-polyfill": "^1.5.0"
   },
   "devDependencies": {
     "babel-core": "^6.22.1",
@@ -52,6 +59,7 @@
     "eslint": "^3.19.0",
     "eslint-config-airbnb-base": "^11.2.0",
     "eslint-plugin-import": "^2.6.1",
+    "extract-text-webpack-plugin": "^2.1.2",
     "fs-extra": "^2.1.2",
     "glob": "^7.1.1",
     "gulp": "^3.9.1",
@@ -59,9 +67,11 @@
     "gulp-clean-css": "^3.4.2",
     "gulp-ignore": "^2.0.2",
     "gulp-mocha": "^2.2.0",
+    "gulp-postcss": "^7.0.1",
     "gulp-rename": "^1.2.2",
     "gulp-require-convert": "0.0.1",
     "gulp-rimraf": "^0.2.0",
+    "gulp-sass": "^3.2.1",
     "gulp-shell": "^0.5.2",
     "gulp-uglify": "^1.5.3",
     "gulp_mcss": "^0.2.0",
@@ -78,12 +88,17 @@
     "np": "^2.16.0",
     "npm-run-all": "^3.1.1",
     "peta": "^1.0.2",
+    "postcss-import": "^11.1.0",
+    "postcss-loader": "^2.1.3",
     "rimraf": "^2.5.4",
     "run-sequence": "^1.2.1",
     "style-loader": "^0.13.1",
     "text-loader": "0.0.1",
-    "webpack": "^1.14.0",
-    "webpack-stream": "^3.2.0",
+    "url-loader": "^1.0.1",
+    "webpack": "^2.6.1",
+    "webpack-cli": "^2.0.14",
+    "webpack-dev-server": "^2.9.1",
+    "webpack-stream": "^4.0.0",
     "yargs": "^6.6.0"
   }
 }
