@@ -9,7 +9,7 @@ const DemoWrap = Regular.extend({
     <div class="m-demowrap">
         <div class="m-codeBlock" r-class="{ {'wrap': !showScript}}" >
           <div class="m-operate-ctn">
-            <kl-button class="m-copyCode" type="tertiary" size="sm" title="复制代码" on-click={this.copy($event)}></kl-button>
+            <kl-button class="m-copyCode" size="sm" title="复制代码" on-click={this.copy($event)}></kl-button>
             <form action="https://codepen.io/pen/define" method="POST" target="_blank" style="display:none;">
               <input type="hidden" name="data" value={codepenJson} />
               <input ref="onlineBtn" type="submit" />
@@ -17,13 +17,13 @@ const DemoWrap = Regular.extend({
           </div>
           <div r-html={htmlTpl} class="figure-outer"></div>
         </div>
-        <div class="m-iconLine" r-class="{ {'watchCode':showScript}}"> 
-          <div class="m-iconOuter" r-class="{ {'revert':showScript}}" on-click={this.toggle()}> 
-            <kl-icon type="caret-down" ></kl-icon>
-          </div> 
+        <div class="m-iconLine" r-class="{ {'watchCode':showScript}}">
+          <div class="m-iconOuter" r-class="{ {'revert':showScript}}" on-click={this.toggle()}>
+            <kl-icon type="solid-arrow-down" />
+          </div>
           <div class="m-iconBtn" r-class="{ {'revertBtn': showScript}}">
             <span class="m-iconTitle" on-click={this.toggle()}>{showScript ? '隐藏代码' : '显示代码'}</span>
-            <kl-button class="m-onlineRun" type="tertiary" size="sm" title="在线运行" on-click={this.online()}>
+            <kl-button class="m-onlineRun" size="sm" title="在线运行" on-click={this.online()}>
             </kl-button>
           </div>
         </div>
@@ -75,8 +75,8 @@ component.$inject(document.body);
       html: html,
       css: css,
       js: js,
-      js_external: 'https://unpkg.com/nek-ui/dist/vendor/regular.min.js;https://unpkg.com/nek-ui/dist/js/nek-ui.min.js',
-      css_external: 'https://unpkg.com/nek-ui/dist/css/nek-ui.red.min.css'
+      js_external: 'https://unpkg.com/nek-ui/dist/vendor/regular.min.js;https://unpkg.com/nek-ui@0.6.2-beta.20/dist/js/nek-ui.min.js',
+      css_external: 'https://unpkg.com/nek-ui@0.6.2-beta.20/dist/css/nek-ui.red.min.css'
     };
     this.data.codepenJson = JSON.stringify(data);
     this.$update();
