@@ -21,7 +21,6 @@ const bowser = require('bowser');
  * @param {string}          [options.data.value]              <=> 文本框的值
  * @param {string}          [options.data.type]               => 文本框的类型, 6种类型：int, float, char，password （email, url暂未实现），
  * @param {string}          [options.data.placeholder]        => 占位符
- * @param {string}          [options.data.state]              <=> 文本框的状态
  * @param {number}          [options.data.maxlength]          => 文本框的最大长度
  * @param {string}          [options.data.unit]               => 单位
  * @param {boolean}         [options.data.clearable=false]    => 是否显示清空图标
@@ -37,7 +36,7 @@ const bowser = require('bowser');
  * @param {number}          [options.data.min]                => 【验证规则】type=int/float时的最小值, type=char时，最小长度
  * @param {number}          [options.data.max]                => 【验证规则】type=int/float时的最大值, type=char时，最大长度
  * @param {string}          [options.data.message]            => 【验证规则】验证失败时，提示的消息
- * @param {string}          [options.data.size]               => 组件大小, sm/md
+ * @param {string}          [options.data.size]               => 组件大小, sm/lg控制整体尺寸，smw/mdw/lgw控制宽度大小
  * @param {number}          [options.data.width]              => 组件宽度
  */
 const KLInput = Component.extend({
@@ -130,7 +129,9 @@ const KLInput = Component.extend({
     let rules = this.data.rules;
 
     rules = rules.filter(rule => (rule.on || '').indexOf(on) >= 0);
-
+    if (!rules.length) {
+      return { success: true };
+    }
     const result = Validation.validate(value, rules);
     if (
       result.firstRule &&
@@ -171,6 +172,27 @@ const KLInput = Component.extend({
       this.data.decimalDigits,
     );
     return value;
+  },
+  /**
+    * @method KLInput#focus 原生focus方法
+    * @param {method} 无 使 input 获取焦点
+  */
+  focus() {
+    this.$refs.input.focus();
+  },
+  /**
+    * @method KLInput#blur 原生blur方法
+    * @param {method} 无 使 input 失去焦点
+  */
+  blur() {
+    this.$refs.input.blur();
+  },
+  /**
+    * @method KLInput#select 原生select方法
+    * @param {method} 无 使 input 选中
+  */
+  select() {
+    this.$refs.input.select();
   },
   _onKeyUp($event) {
     this.validate('keyup');
