@@ -5,6 +5,7 @@
 const scrollIntoViewFn = require('scroll-into-view');
 
 const Component = require('../../../ui-base/component');
+const searchableMixin = require('./mixins/searchable');
 const template = require('./index.html');
 
 /**
@@ -13,18 +14,18 @@ const template = require('./index.html');
  * @param {object}        [options.data]                          => 绑定属性
  * @param {string}        [options.data.class]                    => 补充class
  * @param {array}         [options.data.menus]                    => 菜单数组
- * @param {string}        [options.data.top=60px]               => 菜单style top的值
- * @param {boolean}       [options.data.active=true]              => 默认是否收起
- * @param {string}        [options.data.bodyEl]                => 主内容区body元素的id,当菜单收起时,拉伸bodyEl
+ * @param {string}        [options.data.top=60px]                 => 菜单style top的值
  * @param {boolean}       [options.data.uniqueOpened=true]        => 是否只保持打开一个菜单
  * @param {string}        [options.data.titleKey=title]           => 一级菜单的字段key名
- * @param {string}        [options.data.urlKey=url]             => 菜单结构中的链接key名
- * @param {string}        [options.data.routeKey=route]         => 单页spa应用时,使用
- * @param {string}        [options.data.pageKey=title]          => 二级菜单的字段key名
- * @param {string}        [options.data.childrenKey=children]   => 一级菜单对象下二级菜单数组的key名
+ * @param {string}        [options.data.urlKey=url]               => 菜单结构中的链接key名
+ * @param {string}        [options.data.routeKey=route]           => 单页spa应用时,使用
+ * @param {string}        [options.data.pageKey=title]            => 二级菜单的字段key名
+ * @param {string}        [options.data.childrenKey=children]     => 一级菜单对象下二级菜单数组的key名
+ * @param {string}        [options.data.todoKey=todos]            => 需要显示待办数量时,需要指定key值
  * @param {object}        [options.data.router]                   => 单页应用时, 请将regular-state的manager实例传入
  * @param {string}        [options.data.width]                    => sidebar的宽度设置,默认181px
- * @param {boolean}        [options.data.scrollIntoView=false]          => 是否需要scrollIntoView
+ * @param {string}        [options.data.theme]                    => sidebar的主题颜色，可选light/dark，默认light
+ * @param {boolean}       [options.data.scrollIntoView=false]     => 是否需要scrollIntoView
  */
 const KLSidebar = Component.extend({
   name: 'kl-sidebar',
@@ -39,10 +40,10 @@ const KLSidebar = Component.extend({
       routeKey: 'route',
       pageKey: 'title',
       childrenKey: 'children',
+      todoKey: 'todos',
       top: '60px',
-      active: true,
-      bodyEl: '',
-      width: '181px',
+      width: '216px',
+      theme: 'light',
     });
 
     this.supr();
@@ -50,36 +51,13 @@ const KLSidebar = Component.extend({
   init() {
     const { scrollIntoView } = this.data;
     scrollIntoView && setTimeout(() => {
-      scrollIntoViewFn(document.querySelector('.m-sidebar .m-subMenu .m-menuItem.active'), {
+      scrollIntoViewFn(document.querySelector('.kl-sidebar .kl-sub-menu .kl-menu-item.is-active'), {
         validTarget(target, parentsScrolled) {
-          return parentsScrolled < 2 && target !== window && target.matches('.m-menu');
+          return parentsScrolled < 2 && target !== window && target.matches('.kl-scroll-wrap');
         },
       });
     }, 50);
     this.supr();
-  },
-  initBodyEl() {
-    if (this.data.bodyEl) {
-      this.data.$bodyEl = document.getElementById(this.data.bodyEl);
-      if (this.data.$bodyEl) {
-        this.data.$bodyEl.style.transition = 'left .3s';
-      }
-    }
-  },
-  toggle() {
-    this.initBodyEl();
-
-    this.data.active = !this.data.active;
-    const { width } = this.data;
-    if (this.data.$bodyEl) {
-      this.data.$bodyEl.style.left = this.data.active ? width : '0';
-    }
-
-    /**
-     * @event KLSidebar#toggle 收缩菜单时触发
-     * @property {boolean} active 展开还是收缩
-     */
-    this.$emit('toggle', this.data.active);
   },
   /**
    * @event KLSidebar#menuitem-click 选择某一页时触发
@@ -88,6 +66,24 @@ const KLSidebar = Component.extend({
   onMenuItemClick(e) {
     this.$emit('menuitem-click', e);
   },
+  getTitleTemplate(name, todos) {
+    const title = this.highlight(name);
+    if (todos && !isNaN(todos)) {
+      const count = todos > 99 ? '99+' : todos;
+      return `<span class="kl-sidebar-todowrap"><span class="kl-sidebar-todowrap__title">${title}</span><span class="kl-sidebar-todowrap__badge">${count}</span></span>`;
+    }
+    return title;
+  },
+  getScrollMaxHeight() {
+    const { menuWrap, searchBox } = this.$refs;
+    if (menuWrap && searchBox) {
+      return menuWrap.clientHeight - searchBox.clientHeight;
+    } else if (menuWrap) {
+      return menuWrap.clientHeight;
+    }
+  },
 });
+
+KLSidebar.use(searchableMixin);
 
 module.exports = KLSidebar;
