@@ -11,6 +11,7 @@ const _ = require('../../../ui-base/_');
  * @class KLTabs
  * @extend Component
  * @param {object}        [options.data]                      = 绑定属性
+ * @param {string}        [options.data.type=line]            => 显示类型,card/line
  * @param {object}        [options.data.selected=null]        <=> 当前选择卡
  * @param {string}        [options.data.titleTemplate=null]   @=> 标题模板
  * @param {string}        [options.data.defaultKey=null]      => 默认显示对应 key 的 Tab
@@ -30,7 +31,7 @@ const _ = require('../../../ui-base/_');
 
 const KLTabs = Component.extend({
   name: 'kl-tabs',
-  template,
+  template: _.compressHtml(template),
   config() {
     _.extend(this.data, {
       tabs: [],
@@ -38,6 +39,8 @@ const KLTabs = Component.extend({
       titleTemplate: null,
       offset: 0,
       navStyle: {},
+      type: 'line',
+      crtTabElem: null,
     });
     this.supr();
 
@@ -57,14 +60,14 @@ const KLTabs = Component.extend({
   init() {
     this.supr();
     this._update = this.update.bind(this);
+
     window.addEventListener('resize', this._update);
   },
   events: {
     $init() {
-      const self = this;
       setTimeout(() => {
-        self.update();
-      }, 10);
+        this.update();
+      }, 50);
     },
   },
   update() {
@@ -85,6 +88,7 @@ const KLTabs = Component.extend({
       }
     } else {
       this.data.scrollable = false;
+      this.setOffset(0);
     }
     this.$update();
   },
@@ -95,7 +99,7 @@ const KLTabs = Component.extend({
     navStyle.transform = transform;
     navStyle.msTransform = transform;
     navStyle.webkitTransform = transform;
-    this.update();
+    this.data.scrollable && this.update();
   },
   prev() {
     if (!this.data.scrollable || !this.data.scrollable.prev) {
@@ -128,25 +132,42 @@ const KLTabs = Component.extend({
     const newOffset = navWidth - currentOffset > wrapWidth * 2 ? currentOffset + wrapWidth : (navWidth - wrapWidth);
     this.setOffset(newOffset);
   },
-  select(item) {
+  select(item, e) {
     if (this.data.readonly || this.data.disabled || item.data.disabled) return;
 
     this.data.selected = item;
+    this.data.crtTabElem = e.target;
     /**
      * @event KLTabs#select 选择某一项时触发
      * @property {object} sender 事件发送对象
      * @property {object} selected 当前选择卡
+     * @property {string} key 当前选择卡的key属性
+     * @property {event} e 点击鼠标事件
      */
     this.$emit('select', {
       sender: this,
       selected: item,
       key: item.data.key,
+      e,
     });
   },
   destroy() {
     this.supr();
     window.removeEventListener('resize', this._update);
   },
+});
+
+// eslint-disable-next-line
+KLTabs.directive('active-bar', function(activeBarElem) {
+  this.$watch('selected', (selected) => {
+    if (this.data.type !== 'line' || !selected) { return; }
+    setTimeout(() => {
+      const elem = activeBarElem.parentElement.querySelector('.is-crt');
+
+      activeBarElem.style.width = `${elem.clientWidth}px`;
+      activeBarElem.style.transform = `translateX(${elem.offsetLeft}px)`;
+    }, 100);
+  });
 });
 
 module.exports = KLTabs;
