@@ -12,6 +12,7 @@ const _ = require('../../../ui-base/_');
  * @class KLModal
  * @extend Component
  * @param {object}            [options.data]                      = 绑定属性
+ * @param {string}            [options.data.modalType=提示]             => 对话框类型, 可选参数：`default`、`notice`、`warning`、`error`、`success`
  * @param {string}            [options.data.title=提示]            => 对话框标题
  * @param {string}            [options.data.content]              => 对话框内容
  * @param {string}            [options.data.contentTemplate]      => 对话框内容模板，用于支持复杂内容的自定义。
@@ -28,6 +29,7 @@ const _ = require('../../../ui-base/_');
  * @param {number}            [options.data.maxHeight]            => 内容区域最大高度，超出则显示滚动条
  * @param {string}            [options.data.el]                   => 设置对话框要插入的父级元素，默认为document.body
  * @param {boolean}           [options.data.draggable=false]      => 是否可以拖拽对话框
+ * @param {boolean}           [options.data.fullscreen=false]     => 是否覆盖全屏
  */
 const KLModal = Component.extend({
   name: 'kl-modal',
@@ -36,24 +38,31 @@ const KLModal = Component.extend({
     _.extend(this.data, {
       title: this.$trans('NOTICE'),
       content: '',
+      modalType: '',
       okButton: true,
-      with: 400,
+      width: 400,
       cancelButton: false,
       noClose: false,
       okDisabled: false,
       cancelDisabled: false,
       hasFooter: true,
       isCanClose: true,
+      fullscreen: false,
     });
     this.supr();
   },
 
   init() {
+    // - 禁止Modal的时候后面的内容还可以滚动，很重要的一条！
+    document.body.style.overflow = 'hidden';
     this.supr();
     // 如果不是内嵌组件，则嵌入到data.el或document.body中
     if (this.$root === this) {
       this.$inject(this.data.el || document.body);
     }
+    this.$on('$destroy', () => {
+      document.body.style.overflow = '';
+    });
   },
   /**
      * @method KLModal#close(result) 关闭对话框
@@ -79,7 +88,10 @@ const KLModal = Component.extend({
          * @event KLModal#ok 确定对话框时触发
          */
     this.$emit('ok', event);
-    !this.data.noClose && this.destroy();
+    if (!this.data.noClose) {
+      this.destroy();
+      document.body.style.overflow = '';
+    }
   },
   /**
      * @method KLModal#cancel() 取消对话框
@@ -91,6 +103,8 @@ const KLModal = Component.extend({
          */
     this.$emit('cancel');
     this.destroy();
+    // - 禁止Modal的时候后面的内容还可以滚动，很重要的一条！
+    document.body.style.overflow = '';
   },
   _onDragStart($event) {
     const dialog = $event.proxy;
