@@ -12,14 +12,12 @@ const _ = require('../../../ui-base/_');
  * @class KLProgress
  * @extend Component
  * @param {object}              [options.data]                    = 绑定属性
- * @param {number}              [options.data.percent=36]         => 百分比
- * @param {string/boolean}      [options.data.text=true]          => 在进度条中是否显示百分比。值为`string`时显示该段文字。
- * @param {string}              [options.data.size=null]          => 进度条的尺寸
- * @param {string}              [options.data.state=null]         => 进度条的状态
- * @param {boolean}             [options.data.striped=false]      => 是否显示条纹
- * @param {boolean}             [options.data.active=false]       => 进度条是否为激活状态，当`striped`为`true`时，进度条显示动画
- * @param {boolean}             [options.data.visible=true]       => 是否显示
  * @param {string}              [options.data.class]              => 补充class
+ * @param {number}              [options.data.percent=36]         => 百分比
+ * @param {string}              [options.data.size=null]          => 进度条的尺寸, 可选值sm/lg
+ * @param {boolean}             [options.data.visible=true]       => 是否显示
+ * @param {boolean}             [options.data.insideText=true]    => 进度是否在进度条上显示,size=lg时才生效;
+ * @param {string}             [options.data.color]               => 进度条的颜色
  */
 const KLProgress = Component.extend({
   name: 'kl-progress',
@@ -27,11 +25,7 @@ const KLProgress = Component.extend({
   config() {
     _.extend(this.data, {
       percent: 36,
-      text: true,
-      size: null,
-      state: null,
-      striped: false,
-      active: false,
+      size: '',
     });
     this.supr();
   },
